api/pyproject.toml
[project]
name = "gen_ai_on_aws_app"
version = "0.1.0"
description = "App code for GenAI on AWS"
requires-python = "==3.13.*"
dependencies = [
    "anthropic>=0.49.0",
    "boto3>=1.37.5",
    # FastAPI is a modern web framework for building APIs with Python.
    "fastapi",
    "instructor>=1.7.2",
    "langfuse>=2.59.7",
    "litellm>=1.62.1",
    # Mangum is a library that adapts ASGI applications to AWS Lambda and API Gateway.
    "mangum",
    "pydantic-settings>=2.2.2",
    "supabase>=2.27.1",
]

[dependency-groups]
dev = [
    "devtools>=0.12.2",
    # In development mode, include the FastAPI development server.
    "fastapi[standard]>=0.115",
    "httpx>=0.28.1",
    "ipykernel>=6.29.5",
    "pyright>=1.1.390",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=5.0.0",
    "ruff>=0.10.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-p no:warnings"

[tool.pyright]
include = ["."]
exclude = [
  "build",
  "**/venv",
  "**/.venv",
  "**/env",
  "**/__pycache__",
  "**/.pytest_cache",
  "**/dist",
  "**/.egg-info",
  "**/.mypy_cache",
  "**/.ruff_cache"
]
pythonVersion = "3.13"
typeCheckingMode = "basic"

[tool.ruff]
line-length = 100
target-version = "py313"

[tool.ruff.lint]
select = [
  "E",  # pycodestyle errors
  "W",  # pycodestyle warnings
  "F",  # pyflakes
  "I",  # isort
  "B",  # flake8-bugbear
  "C4", # flake8-comprehensions
  "UP", # pyupgrade
]
ignore = [
  "B008",  # Allow function calls in argument defaults (FastAPI Depends pattern)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
